Revert .gitignore file
For some reason when I cloned the repo I ended up with two files added
/.vs/slnx.sqlite-journal
/.vs/slnx.sqlite

So I used .gitignore to try to ignore them.  But now I see that is showing up a a file change since there was already a .gitignore file for the project.  So this commit is meant to revert that file to it's original unmodified state.